requirements.txt
streamlit
python-dotenv
google-generativeai
gTTS
Pillow